Subscribe to tournament with user_id

+++ src/main/java/com/bezkoder/springjwt/repository/TournamentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TournamentRepository extends JpaRepository<Tournament,  Integer> {
+public interface TournamentRepository extends JpaRepository<Tournament,  Long> {
 }
